make the command interface generic, with a docker implementation
